Regularization optimization update
Select generation = 5 as optimized target trajectories, could be updated by opt_gen in notebook;

The logic of optimization is to choose the smallest regularization when the average correlation across the replicates dropping less then Drop_percent comparing with the neighbooring larger regularization correlation, current Drop_percent is 20 (20%). The searching of the proper regularization starts from the largest regularization.